chore: factor out auth refresh plugin
