+++ clients/web/src/client.ts
@@ -1,14 +1,12 @@
 import {env} from "$env/dynamic/public";
-import {HoudiniClient, RefreshLoginStore} from "$houdini";
+import {HoudiniClient} from "$houdini";
 import {createClient} from "graphql-ws";
 import {subscription} from "$houdini/plugins";
 import {goto} from "$app/navigation";
-import {browser} from "$app/environment";
-import {getAuthCookies, updateAuthCookies} from "$lib/api/auth-cookies";
-import {getExpiryFromJwt} from "./lib/utilities/getExpiryFromJwt";
+import {getAuthCookies} from "$lib/api/auth-cookies";
 import {clearAuthCookies} from "./lib/api/auth-cookies";
 import {redirect} from "@sveltejs/kit";
-import jwtDecode from "jwt-decode";
+import { refreshAuthPlugin } from "./houdini/refresh-auth.plugin";
 
 const getAuthToken = ({
     session,
@@ -61,74 +59,7 @@ export default new HoudiniClient({
         },
     },
     plugins: [
-        () => {
-            return {
-                beforeNetwork: async (ctx, b) => {
-                    const {next} = b;
-                    const {session} = ctx;
-
-                    // TODO: Actually check the thing
-                    if (!session?.access) {
-                        next(ctx);
-                        return;
-                    }
-                    const expAt = getExpiryFromJwt(session?.access);
-
-                    // If there is at least one minute on the token; do nothing
-                    if (expAt.getTime() > new Date().getTime() + 60 * 1000) {
-                        next(ctx);
-                        return;
-                    }
-                    if (ctx.artifact.name === "RefreshLogin") {
-                        console.debug(
-                            "Avoiding infinite loop. Will not try to refresh auth before a refresh auth call.",
-                        );
-                        next(ctx);
-                        return;
-                    }
-
-                    // Otherwise we need to refresh
-                    console.log("Attempting to refresh authentication");
-
-                    if (session.refresh) {
-                        const s = new RefreshLoginStore();
-                        try {
-                            const result = await s.mutate(null, {
-                                metadata: {
-                                    accessTokenOverride: session.refresh,
-                                },
-                                fetch: fetch ?? ctx.fetch,
-                            });
-                            if (!result.data) {
-                                throw new Error(
-                                    result.errors?.map(e => e.message).join("\n") ?? "Refresh Login Response Empty",
-                                );
-                            }
-
-                            if (!ctx.session) ctx.session = {};
-                            ctx.session.access = result.data.refreshLogin.access;
-                            ctx.session.refresh = result.data.refreshLogin.refresh;
-
-                            if (browser) {
-                                updateAuthCookies(result.data.refreshLogin);
-                            } else {
-                                // How do I set cookies here
-                                console.warn(
-                                    "Failed to persist auth cookie updates. Setting cookies in a plugin is not possible?",
-                                );
-                            }
-                            console.debug("Auth has been refreshed");
-                        } catch (e) {
-                            console.warn("Failed to refresh auth token!", e);
-                        }
-                    } else {
-                        console.debug("Skipping Auth Refresh");
-                    }
-
-                    next(ctx);
-                },
-            };
-        },
+        refreshAuthPlugin,
         subscription(ctx =>
             createClient({
                 url: `${env.PUBLIC_GQL_URL}/graphql`,
